login connected to backend. Loading screen. Move from login to home when success

+++ lib/api/user_api.dart
@@ -2,14 +2,14 @@ import 'package:http/http.dart' as http;
 import 'package:gatecheck_frontend/model/user_model.dart';
 import 'package:gatecheck_frontend/api/api_base.dart';
 
-Future<ApiData<GenericUser>> currentUser(http.Client client) {
-  return client.get(api + '/user').then((res) {
+
+Future<ApiData<GenericUser>> currentUser(String token) {
+  return client.get(api + '/user', headers: {"Authorization": "Bearer $token"}).then((res) {
     return parseResponse(res, (json) => GenericUser.fromJson(json));
   });
 }
 
-Future<ApiData<List<Student>>> getStudents(
-    http.Client client, Int32x4List studentIds, bool all) {
+Future<ApiData<List<Student>>> getStudents(Int32x4List studentIds, bool all, String token) {
   String query;
   if (all) {
     query = "all=true";
@@ -17,7 +17,7 @@ Future<ApiData<List<Student>>> getStudents(
     query = "uuid=" + studentIds.map(uuidToString).join(',') + "&all=false";
   }
 
-  return client.get(api + '/user?' + query).then((res) {
+  return client.get(api + '/user?' + query, headers: {"Authorization": "Bearer $token"}).then((res) {
     return parseResponse(res, (value) {
       List<Map<String, dynamic>> json = value.cast<Map<String, dynamic>>();
       List<Student> students = json.map((e) => Student.fromJson(e));
@@ -28,17 +28,17 @@ Future<ApiData<List<Student>>> getStudents(
 
 /// does not immediately delete user, just prompts the server to send him an email to delete himself
 
-Future<ApiData<void>> deleteUser(http.Client client) {
-  return client.delete(api + '/user').then(emptyResponseData);
+Future<ApiData<void>> deleteUser(String token) {
+  return client.delete(api + '/user', headers: {"Authorization": "Bearer $token"}).then(emptyResponseData);
 }
 
-Future<ApiData<void>> updateCurrentUser(http.Client client, GenericUser user) {
-  return client.put(api + '/user', body: user.toJson()).then(emptyResponseData);
+Future<ApiData<void>> updateCurrentUser(GenericUser user, String token) {
+  return client.put(api + '/user', body: user.toJson(), headers: {"Authorization": "Bearer $token"}).then(emptyResponseData);
 }
 
-Future<ApiData<void>> updateSpecificUser(http.Client client, GenericUser user) {
+Future<ApiData<void>> updateSpecificUser(GenericUser user, String token) {
   var data = user.toJson();
   return client
-      .put(api + '/user/' + data['id'], body: data)
+      .put(api + '/user/' + data['id'], body: data, headers: {"Authorization": "Bearer $token"})
       .then(emptyResponseData);
 }
